Prevent overflow by pre-casting to int64_t
When converting to ticks it is possible for the first value to overflow.
We can prevent this overflow if we cast the first value to int64_t
before multiplying. This will cause the compiler to use int64_t as the
base type for the operations.

+++ src/dvdnav.c
@@ -203,12 +203,12 @@ int64_t dvdnav_convert_time(dvd_time_t *time) {
   int64_t result;
   int64_t frames;
 
-  result  = (time->hour    >> 4  ) * 10 * 60 * 60 * 90000;
-  result += (time->hour    & 0x0f)      * 60 * 60 * 90000;
-  result += (time->minute  >> 4  )      * 10 * 60 * 90000;
-  result += (time->minute  & 0x0f)           * 60 * 90000;
-  result += (time->second  >> 4  )           * 10 * 90000;
-  result += (time->second  & 0x0f)                * 90000;
+  result  = ((int64_t)(time->hour    >> 4  )) * 10 * 60 * 60 * 90000;
+  result += ((int64_t)(time->hour    & 0x0f))      * 60 * 60 * 90000;
+  result += ((int64_t)(time->minute  >> 4  ))      * 10 * 60 * 90000;
+  result += ((int64_t)(time->minute  & 0x0f))           * 60 * 90000;
+  result += ((int64_t)(time->second  >> 4  ))           * 10 * 90000;
+  result += ((int64_t)(time->second  & 0x0f))                * 90000;
 
   frames  = ((time->frame_u & 0x30) >> 4) * 10;
   frames += ((time->frame_u & 0x0f)     )     ;
